♻️ Chatコンポーネントをリファクタ
