+++ src/hooks/useChat.tsx
@@ -1,20 +1,40 @@
-import { useCallback, useRef, useState } from 'react'
+import { db } from 'firebase/clientApp'
+import { useCallback, useEffect, useRef, useState } from 'react'
 
 type ChatTable = {
   message: string
   id: string
 }[]
 
-export const useChat = () => {
+export const chatCollectionRef = (id?: string) => db.collection('contents').doc(id).collection('thread')
+
+export const useChat = (roomId?: string) => {
   const [chatText, setChatText] = useState('')
   const [isChatOpen, isNotChatOpen] = useState(false)
   const [chatTable, setChatTable] = useState<ChatTable>([])
 
-  const scrollRef = useRef<HTMLDivElement>(null)
+  //チャットデータ取得
+  useEffect(() => {
+    const unsub = chatCollectionRef(roomId)
+      .orderBy('createdAt')
+      .onSnapshot((snapshot) => {
+        setChatTable(snapshot.docs.map((doc) => ({ message: doc.data().message, id: doc.id })))
+      })
+    return () => unsub()
+  }, [])
 
-  const scrollChatList = useCallback(() => {
-    scrollRef?.current?.scrollIntoView()
-  }, [scrollRef])
+  //チャット送信
+  const hundleAdd = useCallback(
+    async (chatText) => {
+      await chatCollectionRef(roomId).add({
+        message: chatText,
+        createdAt: JSON.stringify(new Date()),
+      })
+      setChatText('')
+      scrollChatList()
+    },
+    [chatText]
+  )
 
   const hundleText = useCallback(
     (e) => {
@@ -27,15 +47,21 @@ export const useChat = () => {
     isNotChatOpen(!isChatOpen)
   }, [isChatOpen])
 
+  //スクロール制御
+  const scrollRef = useRef<HTMLDivElement>(null)
+  const scrollChatList = useCallback(() => {
+    scrollRef?.current?.scrollIntoView()
+  }, [scrollRef])
+
   return {
     chatText,
     hundleText,
-    setChatText,
     isChatOpen,
     hundleChat,
     chatTable,
     setChatTable,
     scrollRef,
     scrollChatList,
+    hundleAdd,
   }
 }
